Removes beginning of dissolution data from fit
This way it's not trying to fit to data from a prior dissolution reaction, that isn't included in the current model. Doesn't have a big effect since the surface coverage is approximated as zero there. Though I kind of wonder what would happen if we had modeled two potential sweeps back-to-back - does some theta_2 stick around after the first sweep?